Updated User.py to include concept of static
user_count is a static variable that tracks the number of users created.
reset_count() is a class method that resets the user count. (This will cause a collision in user_ids - a BAD THING)

double_count() is a static method that doubles the value in user_count (this will create a jump in subsequently generated user_ids)