Added exercise code from Exception handling exercise. Original version includes no exception handling and is vulnerable to issues where a line in the file is malformatted (insufficient number of elements included (IndexError) or text included where number is expected (ValueError)) or the file cannot be found (FileNotFoundError)
general exception handled application includes handling for file not found error, plus a general except block to handle any IndexErrors in a single place. This version is less specific and means you can't recover anything from that file (it drops straight to ask the user to enter a new file name), but means the code is neater, and where information in the file is linked / relies on other parts, this can be the only sensible way to deal with it.